+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -1,5 +1,6 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.path.domain;
 
+import nextstep.subway.line.domain.Line;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.domain.Station;
 import org.springframework.stereotype.Component;
